🚚 feat: Restructure project with modules directory
- Create a new 'modules' directory to organize project modules
- Move existing modules into the new 'modules' directory
- Update import paths to reflect the new structure
