+++ src/modules/app/app.module.ts
@@ -6,11 +6,11 @@ import {
 } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { LoggerMiddleware } from './middlewares/logger.middleware';
+import { LoggerMiddleware } from '../../middlewares/logger.middleware';
 import { ConfigModule } from '@nestjs/config';
-import { PrismaModule } from './prisma/prisma.module';
-import config from './common/configs/config';
-import { AuthModule } from './auth/auth.module';
+import { PrismaModule } from '../prisma/prisma.module';
+import config from '../../common/configs/config';
+import { AuthModule } from '../auth/auth.module';
 
 @Module({
   imports: [
